Add files via upload
Yield Generator Cicada: An attempt to optimize Yield Generator Oscillator for profit, money velocity, and information leakage avoidance.

It does a regression analysis on past transactions, using recent transactions based on randomly chosen amount ranges and time periods, in order to determine the power law of the form earnings = C+A*pow(amount,B) with the highest correlation, and reduces the coefficient A further based on an average amount of transactions per offer range in order to partially offset price increments so that unpopular amount ranges will have offers lower than average coinjoin fee just as popular amount ranges have higher than average coinjoin fee. Then it picks a new set of amount ranges and time periods and generates offers with some statistical noise based on uncertainties in A and B. It repeats that process every hour.

I expect it has a division by zero error in calculating correlations if you run it on a new wallet but I haven't tested it that way. It definitely