Update yield-generator-cicada.py
This does a bunch of statistics, controlled randomization, and sampling using randomly chosen prime-numbered time ranges to arrive at rational pricing that attempts to stay as random as the underlying data and range of likely realistic preferences for a maker. It is still based on the logic of the oscillator yield generator - popular amount ranges are still priced upwards - but the underlying offers and price increments are calculated in a very involved way to balance rational pricing (which itself requires randomness to charge impatient parties more than patient parties) based on private transaction data, with obscuring of that private transaction data. Users can easily make the utility function non-random by specifying specific weights in place of the random.random() calls.