Update yield-generator-cicada.py
We aim to shape the randomness of offers by cutting off the tails of our gaussian distributions - and also making the deviation of the exponent dependent on the deviation of the coefficient. The idea is to get rid of long tails while also putting some bias towards being "on sale" to counteract the sampling bias that only accepted offers are sampled - and sometimes including a tip for a final price that might have been unattractive to begin with.

I regret to say that this may be my final commit to this script and m final contribution to this project for a while. I have tentatively ceased doing CoinJoins on JoinMarket today following the hack on Bitfinex because I am spooked about the idea of a thief potentially tumbling coins through JoinMarket - however unlikely that may be. Even if I should later choose to resume, I have little incentive to improve this project as the most needful improvements are outside the scope of my expertise - and the little pricing tricks I enjoy improvising are more profitably applied to trading and arbitrage. So I leave you with this last commit which is the last yield generator I used.